+++ xb8_docsis_stats/xb8-stats.py
@@ -4,118 +4,58 @@ __author__ = "Aaron Turner <synfinatic at gmail dot com"
 
 import argparse
 from bs4 import BeautifulSoup
-from enum import Enum
 import logging
 import graphyte
 import requests
 import time
-from typing import Dict
+from timeit import default_timer as timer
+
 
 from lib.args import EnvDefault
 from lib.channel import Tables
+from lib.logging import LogLevel
+from lib.init_proc import InitializationProcedure
+from lib.http import XBRConfig, fetch_network_stats_page
 
 log = None
 
 
-class LogLevel(Enum):
-    error = 'error'
-    info = 'info'
-    warn = 'warn'
-    debug = 'debug'
-
-    def __str__(self):
-        return self.value
-
-    @property
-    def level(self) -> int:
-        levels = {
-            'debug': logging.DEBUG,
-            'warn': logging.WARNING,
-            'info': logging.INFO,
-            'error': logging.ERROR,
-            'critical': logging.CRITICAL,
-        }
-        return levels[self.value]
-
-
-def login(args) -> Dict[str, str]:
-    page = requests.post(f'http://{args.modem}/check.jst',
-        data={
-            'username': args.username,
-            'password': args.password,
-        })
-    page.raise_for_status()
-    if 'Access denied!' in str(page.content):
-        raise ValueError('Invalid username/password')
-
-    return page.cookies
-
-
-def fetch_stats(args) -> Tables:
-    cookie_jar = login(args)
-
-    page = requests.get(f'http://{args.modem}/network_setup.jst',
-                        cookies=cookie_jar)
-    page.raise_for_status()
-    soup = BeautifulSoup(page.content, "html.parser")
-
-    content = soup.find(id='content')
-    modules = content.find_all("div", class_="module")
-
-    tables = Tables()
-
-    midx = 0
-    for i, m in enumerate(modules):
-        # seems to be more <div class="module"> than I expect??
-        table = m.find("table", class_="data")
-        if table is None:
-            continue
-
-        mtype = Tables.fields[midx]
-        midx += 1
-
-        # process table body
-        body = table.find("tbody")
-        for ridx, row in enumerate(body.find_all("tr")):
-            for cidx, column in enumerate(row.find_all("td")):
-                val = column.text.strip()
-                if ridx == 0:
-                    tables.new(mtype)
-                try:
-                    tables.add(mtype, cidx, ridx, val)
-                except ValueError as e:
-                    log.error(
-                        f'mtype = {mtype}, cidx = {cidx}, '
-                        f'ridx = {ridx}, val = {val}')
-                    raise e
-    return tables
-
-
-def submit_metrics(args, tables: Tables):
+def submit_metrics(args, tables: Tables, init_proc: InitializationProcedure):
     host, port = args.graphite.split(':')
     sender = graphyte.Sender(host, port=port, raise_send_errors=True)
     i = 0
+    for metric, value in init_proc.metrics.items():
+        name = '.'.join([args.prefix, 'InitProcedure', metric])
+        sender.send(name, value.value)
+        i += 1
+
     for k in ['Downstream', 'Upstream']:
         for table in getattr(tables, k):
             for metric in table.metrics:
                 name = '.'.join([args.prefix, metric.Name])
-                i += 1
                 sender.send(name, metric.Value)
+                i += 1
+
     sender.stop()
     log.info('sent %d metrics', i)
 
 
-def loop(args):
+def loop(config: XBRConfig, args):
     tables = None
     try:
-        tables = fetch_stats(args)
-    except requests.exceptions.ConnectTimeout as e:
+        tables = Tables()
+        init_proc = InitializationProcedure()
+        page = fetch_network_stats_page(config)
+        tables.load(page)
+        init_proc.load(page)
+    except requests.exceptions.RequestException as e:
         log.error("Unable to poll modem: %s", e)
-    tables.map_channels()
+    # let ValueError() to up the stack as it is fatal
     if args.dump_json:
         print(tables.to_json())
+        print(init_proc.to_json())
     else:
-        submit_metrics(args, tables)
+        submit_metrics(args, tables, init_proc)
 
 
 def main():
@@ -128,6 +68,7 @@ def main():
         description=__description__,
     )
 
+    action_group = parser.add_mutually_exclusive_group(required=True)
     parser.add_argument(
         '--modem', type=str, default='10.0.0.1',
         help='IP or hostname of the modem (%(default)s)')
@@ -138,8 +79,11 @@ def main():
         '--password', '-p', action=EnvDefault,
         envvar='XB8_PASSWORD', required=True,
         help='Password to use to login to the modem')
-    parser.add_argument(
-        '--graphite', type=str, required=True,
+    action_group.add_argument(
+        '--dump-json', action='store_true',
+        help='Dumps the stats as JSON instead')
+    action_group.add_argument(
+        '--graphite', type=str,
         help='host:port of the graphite submission server')
     parser.add_argument(
         '--prefix', type=str, default='modem',
@@ -148,20 +92,28 @@ def main():
         '--interval', type=int, default=0,
         help='Number of seconds between each poll cycle')
     parser.add_argument(
-        '--dump-json', action='store_true',
-        help='Dumps the stats as JSON instead')
-    parser.add_argument(
         '--log-level', type=LogLevel, choices=list(LogLevel),
         default='info', help='Log level (%(default)s)')
 
     args = parser.parse_args()
     log.setLevel(args.log_level.level)
 
-    loop(args)
+    config = XBRConfig(
+        Hostname=args.modem,
+        Username=args.username,
+        Password=args.password,
+    )
+
+    start = timer()
+    loop(config, args)
+    end = timer()
     if args.interval:
         while True:
-            time.sleep(args.interval)
-            loop(args)
+            # adjust our sleep time to compensate for how slow the modem is
+            time.sleep(args.interval - (end - start))
+            start = timer()
+            loop(config, args)
+            end = timer()
 
 
 if __name__ == "__main__":